fix: improve performance for latest episodes page
