+++ src/app/pages/podcasts/latest-episodes.tsx
@@ -1,27 +1,24 @@
 import { useQuery } from '@tanstack/react-query'
-import { useVirtualizer } from '@tanstack/react-virtual'
-import { useEffect, useMemo, useRef } from 'react'
+import { memo, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ShadowHeader } from '@/app/components/album/shadow-header'
 import { LatestEpisodesFallback } from '@/app/components/fallbacks/podcast-fallbacks'
 import { HeaderTitle } from '@/app/components/header-title'
 import ListWrapper from '@/app/components/list-wrapper'
-import { EpisodeCard } from '@/app/components/podcasts/episode-card'
 import { FeaturedEpisodeCard } from '@/app/components/podcasts/featured-episode-card'
+import { LatestEpisodesList } from '@/app/components/podcasts/latest-episodes-list'
 import ErrorPage from '@/app/pages/error-page'
 import { podcasts } from '@/service/podcasts'
 import { queryKeys } from '@/utils/queryKeys'
-import { getMainScrollElement } from '@/utils/scrollPageToTop'
 
 const episodesToFeature = 5
 
+const MemoShadowHeader = memo(ShadowHeader)
+const MemoHeaderTitle = memo(HeaderTitle)
+const MemoFeaturedEpisodeCard = memo(FeaturedEpisodeCard)
+
 export default function LatestEpisodes() {
   const { t } = useTranslation()
-  const scrollDivRef = useRef<HTMLDivElement | null>(null)
-
-  useEffect(() => {
-    scrollDivRef.current = getMainScrollElement()
-  }, [])
 
   const { data, isFetched, isFetching, isLoading } = useQuery({
     queryKey: [queryKeys.episode.latest],
@@ -36,56 +33,25 @@ export default function LatestEpisodes() {
     return data ? data.slice(0, episodesToFeature) : []
   }, [data])
 
-  const virtualizer = useVirtualizer({
-    count: episodes.length,
-    getScrollElement: () => scrollDivRef.current,
-    estimateSize: () => 124,
-    overscan: 5,
-  })
-
   if (isLoading || isFetching) return <LatestEpisodesFallback />
   if (isFetched && !data) {
     return <ErrorPage />
   }
 
-  const items = virtualizer.getVirtualItems()
-
   return (
     <div className="w-full">
-      <ShadowHeader>
-        <HeaderTitle title={t('podcasts.form.latestEpisodes')} />
-      </ShadowHeader>
+      <MemoShadowHeader>
+        <MemoHeaderTitle title={t('podcasts.form.latestEpisodes')} />
+      </MemoShadowHeader>
 
       <ListWrapper className="px-4 pt-[--shadow-header-distance]">
         <div className="grid grid-cols-5 gap-4 px-4 mb-6">
           {featuredEpisodes.map((episode) => (
-            <FeaturedEpisodeCard key={episode.id} episode={episode} />
+            <MemoFeaturedEpisodeCard key={episode.id} episode={episode} />
           ))}
         </div>
 
-        <div
-          style={{
-            height: virtualizer.getTotalSize(),
-            position: 'relative',
-          }}
-        >
-          {items.map((virtualRow) => {
-            const episode = episodes[virtualRow.index]
-
-            return (
-              <EpisodeCard
-                episode={episode}
-                key={virtualRow.index}
-                latest={true}
-                style={{
-                  position: 'absolute',
-                  top: virtualRow.start,
-                  width: '100%',
-                }}
-              />
-            )
-          })}
-        </div>
+        <LatestEpisodesList episodes={episodes} />
       </ListWrapper>
     </div>
   )
